requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
tensorflow
openai
python-dotenv
```

### 2. **`Procfile`** (optional pero recommended)
```
web: uvicorn main:app --host 0.0.0.0 --port $PORT
```

**Note:** Palitan mo yung `main` kung iba yung filename ng Python file mo (e.g., kung `app.py` yung file mo, gawin mong `app:app`)

---

## ⚠️ IMPORTANT: Security Issue!

**HUWAG I-PUSH YUNG `.env` FILE SA GITHUB!** 

Kitang-kita ko yung OpenAI API key mo dyan! 🚨

### Gawin mo to NOW:

1. **Add `.gitignore` file:**
```
.env
__pycache__/
*.pyc
*.pyo
*.pyd
.Python
venv/
env/
```

2. **Revoke yung API key mo ASAP:**
   - Go to https://platform.openai.com/api-keys
   - Delete yung exposed key
   - Create new key

3. **For Render deployment, yung API key ilalagay mo sa Environment Variables**, NOT sa code

---

## File Structure mo dapat ganito:
```
dermaware_backend/
├── main.py (or app.py - yung FastAPI code mo)
├── model_unquant.tflite
├── labels.txt
├── requirements.txt  ← CREATE THIS
├── Procfile           ← CREATE THIS
├── .env               ← HUWAG I-PUSH SA GITHUB
└── .gitignore         ← CREATE THIS